+++ tests/cases/static_nglob_recursive/plan.py
@@ -1,0 +1,6 @@
+#!/usr/bin/env python
+from stepup.core.api import glob
+
+m = glob("data/**")
+for path in m:
+    print(path)